fix: Allow job sync with cluster tokens
