+++ control-plane/src/modules/management.ts
@@ -269,6 +269,8 @@ export const getClusterSettings = async (clusterId: string) => {
   return settings;
 };
 
+export const ACCESS_POINT_TOKEN_PREFIX = "sk_ap_";
+
 export const upsertAccessPointForCluster = async ({
   clusterId,
   name,
@@ -278,7 +280,7 @@ export const upsertAccessPointForCluster = async ({
   name: string;
   allowedServices: string;
 }) => {
-  const token = `sk_ap_${crypto.randomBytes(16).toString("hex")}`;
+  const token = `${ACCESS_POINT_TOKEN_PREFIX}${crypto.randomBytes(16).toString("hex")}`;
 
   const [accessPoint] = await data.db
     .insert(data.clusterAccessPoints)
